deploy/requirements.txt
aws-cdk-lib==2.32.1
constructs>=10.0.0,<11.0.0
constructs>=10.0.0,<11.0.0
sagemaker>=2.140.0
transformers==4.26.1
datasets[s3]==2.10.1
boto3